+++ 01.Grafos/archivos_ej1/01.csv
@@ -1,0 +1,6 @@
+1,0,0,0,0,0
+0,1,0,0,0,0
+0,0,1,0,0,0
+0,0,0,1,0,0
+0,0,0,0,1,0
+0,0,0,0,0,1
