+++ 01.Grafos/archivos_ej1/04.csv
@@ -1,0 +1,6 @@
+0,0,1,0,0,1
+1,1,1,0,0,0
+0,1,1,0,0,0
+0,0,0,1,0,0
+0,0,0,0,0,0
+1,0,0,0,0,0
