+++ 01.Grafos/archivos_ej3/esDivisorDe-200.json
@@ -1,0 +1,105 @@
+
+{
+	"P": ["1","2","3","4","5","6","7","8","9","10","11","12","13","14","15","16","17","18","19","20","21","22","23","24","25","26","27","28","29","30","31","32","33","34","35","36","37","38","39","40","41","42","43","44","45","46","47","48","49","50","51","52","53","54","55","56","57","58","59","60","61","62","63","64","65","66","67","68","69","70","71","72","73","74","75","76","77","78","79","80","81","82","83","84","85","86","87","88","89","90","91","92","93","94","95","96","97","98","99","100","101","102","103","104","105","106","107","108","109","110","111","112","113","114","115","116","117","118","119","120","121","122","123","124","125","126","127","128","129","130","131","132","133","134","135","136","137","138","139","140","141","142","143","144","145","146","147","148","149","150","151","152","153","154","155","156","157","158","159","160","161","162","163","164","165","166","167","168","169","170","171","172","173","174","175","176","177","178","179","180","181","182","183","184","185","186","187","188","189","190","191","192","193","194","195","196","197","198","199"],
+	"E": {
+	"1": ["2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50", "51", "52", "53", "54", "55", "56", "57", "58", "59", "60", "61", "62", "63", "64", "65", "66", "67", "68", "69", "70", "71", "72", "73", "74", "75", "76", "77", "78", "79", "80", "81", "82", "83", "84", "85", "86", "87", "88", "89", "90", "91", "92", "93", "94", "95", "96", "97", "98", "99", "100", "101", "102", "103", "104", "105", "106", "107", "108", "109", "110", "111", "112", "113", "114", "115", "116", "117", "118", "119", "120", "121", "122", "123", "124", "125", "126", "127", "128", "129", "130", "131", "132", "133", "134", "135", "136", "137", "138", "139", "140", "141", "142", "143", "144", "145", "146", "147", "148", "149", "150", "151", "152", "153", "154", "155", "156", "157", "158", "159", "160", "161", "162", "163", "164", "165", "166", "167", "168", "169", "170", "171", "172", "173", "174", "175", "176", "177", "178", "179", "180", "181", "182", "183", "184", "185", "186", "187", "188", "189", "190", "191", "192", "193", "194", "195", "196", "197", "198", "199"],
+	"2": ["4", "6", "8", "10", "12", "14", "16", "18", "20", "22", "24", "26", "28", "30", "32", "34", "36", "38", "40", "42", "44", "46", "48", "50", "52", "54", "56", "58", "60", "62", "64", "66", "68", "70", "72", "74", "76", "78", "80", "82", "84", "86", "88", "90", "92", "94", "96", "98", "100", "102", "104", "106", "108", "110", "112", "114", "116", "118", "120", "122", "124", "126", "128", "130", "132", "134", "136", "138", "140", "142", "144", "146", "148", "150", "152", "154", "156", "158", "160", "162", "164", "166", "168", "170", "172", "174", "176", "178", "180", "182", "184", "186", "188", "190", "192", "194", "196", "198"],
+	"3": ["6", "9", "12", "15", "18", "21", "24", "27", "30", "33", "36", "39", "42", "45", "48", "51", "54", "57", "60", "63", "66", "69", "72", "75", "78", "81", "84", "87", "90", "93", "96", "99", "102", "105", "108", "111", "114", "117", "120", "123", "126", "129", "132", "135", "138", "141", "144", "147", "150", "153", "156", "159", "162", "165", "168", "171", "174", "177", "180", "183", "186", "189", "192", "195", "198"],
+	"4": ["8", "12", "16", "20", "24", "28", "32", "36", "40", "44", "48", "52", "56", "60", "64", "68", "72", "76", "80", "84", "88", "92", "96", "100", "104", "108", "112", "116", "120", "124", "128", "132", "136", "140", "144", "148", "152", "156", "160", "164", "168", "172", "176", "180", "184", "188", "192", "196"],
+	"5": ["10", "15", "20", "25", "30", "35", "40", "45", "50", "55", "60", "65", "70", "75", "80", "85", "90", "95", "100", "105", "110", "115", "120", "125", "130", "135", "140", "145", "150", "155", "160", "165", "170", "175", "180", "185", "190", "195"],
+	"6": ["12", "18", "24", "30", "36", "42", "48", "54", "60", "66", "72", "78", "84", "90", "96", "102", "108", "114", "120", "126", "132", "138", "144", "150", "156", "162", "168", "174", "180", "186", "192", "198"],
+	"7": ["14", "21", "28", "35", "42", "49", "56", "63", "70", "77", "84", "91", "98", "105", "112", "119", "126", "133", "140", "147", "154", "161", "168", "175", "182", "189", "196"],
+	"8": ["16", "24", "32", "40", "48", "56", "64", "72", "80", "88", "96", "104", "112", "120", "128", "136", "144", "152", "160", "168", "176", "184", "192"],
+	"9": ["18", "27", "36", "45", "54", "63", "72", "81", "90", "99", "108", "117", "126", "135", "144", "153", "162", "171", "180", "189", "198"],
+	"10": ["20", "30", "40", "50", "60", "70", "80", "90", "100", "110", "120", "130", "140", "150", "160", "170", "180", "190"],
+	"11": ["22", "33", "44", "55", "66", "77", "88", "99", "110", "121", "132", "143", "154", "165", "176", "187", "198"],
+	"12": ["24", "36", "48", "60", "72", "84", "96", "108", "120", "132", "144", "156", "168", "180", "192"],
+	"13": ["26", "39", "52", "65", "78", "91", "104", "117", "130", "143", "156", "169", "182", "195"],
+	"14": ["28", "42", "56", "70", "84", "98", "112", "126", "140", "154", "168", "182", "196"],
+	"15": ["30", "45", "60", "75", "90", "105", "120", "135", "150", "165", "180", "195"],
+	"16": ["32", "48", "64", "80", "96", "112", "128", "144", "160", "176", "192"],
+	"17": ["34", "51", "68", "85", "102", "119", "136", "153", "170", "187"],
+	"18": ["36", "54", "72", "90", "108", "126", "144", "162", "180", "198"],
+	"19": ["38", "57", "76", "95", "114", "133", "152", "171", "190"],
+	"20": ["40", "60", "80", "100", "120", "140", "160", "180"],
+	"21": ["42", "63", "84", "105", "126", "147", "168", "189"],
+	"22": ["44", "66", "88", "110", "132", "154", "176", "198"],
+	"23": ["46", "69", "92", "115", "138", "161", "184"],
+	"24": ["48", "72", "96", "120", "144", "168", "192"],
+	"25": ["50", "75", "100", "125", "150", "175"],
+	"26": ["52", "78", "104", "130", "156", "182"],
+	"27": ["54", "81", "108", "135", "162", "189"],
+	"28": ["56", "84", "112", "140", "168", "196"],
+	"29": ["58", "87", "116", "145", "174"],
+	"30": ["60", "90", "120", "150", "180"],
+	"31": ["62", "93", "124", "155", "186"],
+	"32": ["64", "96", "128", "160", "192"],
+	"33": ["66", "99", "132", "165", "198"],
+	"34": ["68", "102", "136", "170"],
+	"35": ["70", "105", "140", "175"],
+	"36": ["72", "108", "144", "180"],
+	"37": ["74", "111", "148", "185"],
+	"38": ["76", "114", "152", "190"],
+	"39": ["78", "117", "156", "195"],
+	"40": ["80", "120", "160"],
+	"41": ["82", "123", "164"],
+	"42": ["84", "126", "168"],
+	"43": ["86", "129", "172"],
+	"44": ["88", "132", "176"],
+	"45": ["90", "135", "180"],
+	"46": ["92", "138", "184"],
+	"47": ["94", "141", "188"],
+	"48": ["96", "144", "192"],
+	"49": ["98", "147", "196"],
+	"50": ["100", "150"],
+	"51": ["102", "153"],
+	"52": ["104", "156"],
+	"53": ["106", "159"],
+	"54": ["108", "162"],
+	"55": ["110", "165"],
+	"56": ["112", "168"],
+	"57": ["114", "171"],
+	"58": ["116", "174"],
+	"59": ["118", "177"],
+	"60": ["120", "180"],
+	"61": ["122", "183"],
+	"62": ["124", "186"],
+	"63": ["126", "189"],
+	"64": ["128", "192"],
+	"65": ["130", "195"],
+	"66": ["132", "198"],
+	"67": ["134"],
+	"68": ["136"],
+	"69": ["138"],
+	"70": ["140"],
+	"71": ["142"],
+	"72": ["144"],
+	"73": ["146"],
+	"74": ["148"],
+	"75": ["150"],
+	"76": ["152"],
+	"77": ["154"],
+	"78": ["156"],
+	"79": ["158"],
+	"80": ["160"],
+	"81": ["162"],
+	"82": ["164"],
+	"83": ["166"],
+	"84": ["168"],
+	"85": ["170"],
+	"86": ["172"],
+	"87": ["174"],
+	"88": ["176"],
+	"89": ["178"],
+	"90": ["180"],
+	"91": ["182"],
+	"92": ["184"],
+	"93": ["186"],
+	"94": ["188"],
+	"95": ["190"],
+	"96": ["192"],
+	"97": ["194"],
+	"98": ["196"],
+	"99": ["198"]
+	}
+}
